new title The Killer, bug fixes and improvements, decimals to vote, smooth average, several winners

+++ src/main/java/com/ppot14/futbol7/DBConnector.java
@@ -171,7 +171,7 @@ public class DBConnector {
 			e.printStackTrace();
 		}
 	}
-	public static Boolean addPunctuation(String voter, String voted, Integer score, String comment, Long dateL, String season){
+	public static Boolean addPunctuation(String voter, String voted, Double score, String comment, Long dateL, String season){
 		String date = formatter.format(new Date(dateL));
 		try {
 			MongoCollection<Document> configCollection;
